requirements.txt
alabaster==0.7.12
astroid==2.1.0
autopep8==1.4.3
Babel==2.6.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
codecov==2.0.15
coverage==4.5.2
doc8==0.8.0
docutils==0.14
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
imagesize==1.1.0
isort==4.3.4
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
packaging==19.0
pbr==5.1.2
peewee==3.8.1
pkginfo==1.5.0.1
pycodestyle==2.5.0
pyfiglet==0.8.post1
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.3.1
python-decouple==3.1
pytz==2018.7
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.8.0
restructuredtext-lint==1.2.2
six==1.12.0
smmap2==2.0.5
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
stevedore==1.30.0
tqdm==4.29.1
twine==1.12.1
typed-ast==1.2.0
urllib3==1.24.1
webencodings==0.5.1
wrapt==1.11.1
